UITrigger::Trigger: std::fabs promoted the anim-length divide to double
96.87 -> 98.2 canonical (472 B).  std::fabs takes and returns DOUBLE, so
f4 = std::fabs(mStart - mEnd) / f4 compiled to a double divide plus a
rounding step -- fdiv + frsp -- where the image has a single
fdivs f0, f13, f0 at 827B26E8.  The fabs instruction itself was never the
tell: PPC fabs is register-width and both sides emit it identically; only
the divide's width differed.  Switched both std::fabs sites in the function
to fabsf (the other one, on StartFrame()-EndFrame(), is fsubs + fabs in the
image too).

This is a real precision divergence, not just a byte count: the image
computes (mStart - mEnd) / scale entirely in single precision.

Residual 35 rows: an r27<->r28 / f29<->f30 permutation and the loop base --
the image addresses the Anim fields straight off the list node (0x14, 0x28,
0x30..0x3c) where we materialise addi r29, r30, 0x8 and use displacements
8 lower.

+++ src/system/ui/UITrigger.cpp
@@ -70,10 +70,16 @@ void UITrigger::Trigger() {
                     if (!f4) {
                         f4 = 1.0f;
                     }
-                    f4 = std::fabs(curAnim.mStart - curAnim.mEnd) / f4;
+                    // fabsf, not std::fabs: std::fabs takes and returns
+                    // DOUBLE, so the divide below came out as `fdiv` + `frsp`
+                    // where the image has a single `fdivs f0, f13, f0`
+                    // (827B26E8).  Both sides emit the same `fabs` instruction
+                    // -- it is the divide's precision that the double round
+                    // trip changed.
+                    f4 = fabsf(curAnim.mStart - curAnim.mEnd) / f4;
                 }
             } else {
-                f4 = std::fabs(curAnim.mAnim->StartFrame() - curAnim.mAnim->EndFrame());
+                f4 = fabsf(curAnim.mAnim->StartFrame() - curAnim.mAnim->EndFrame());
             }
             MaxEq(mEndTime, (curAnim.mDelay * 30.0f + f4) / 30.0f);
         }
